Support running debug tests with java
This CL allows to run the debug tests with java. It makes it easier
to compare debugging behavior between java and art.

In case of expected behavior difference, the test needs to support
different checks. An example in this CL is the ExceptionTest where
stepping through an exception leads to different locations depending
on the selected runtime.

We keep running these tests with art by default. In the future, we
may want to run these tests in both modes.

Change-Id: I99ccf52a30b8fe4d3e2f47516588239c5ac0c1a3