Tweak handling of arguments when argument register reuse is disabled.
If there are multiple uses of an argument value, split it right after
its definition instead of at the first use. This makes it more likely
that arguments get in useful registers initially and reduces the
amount of argument moves needed.

This reduces the GMSCore v10 debug build size by 20kB and reduces
a simple debug dex file size from 15kB to 13kB.

R=zerny@google.com

Bug: 38375244
Change-Id: Ibfe45500a87c73a91dad6bbcfdb048dd590e5efb