Improve argument register selection and debug info handling.
When allocating a register for an argument value split with no register
constraint, just use the original argument register.

For debug info, always use the original argument register for an
argument value. When we have splits we know that we have disallowed
argument register reuse and therefore the argument value is always
available in the original argument register.

R=zerny@google.com

Bug: 62186207
Change-Id: I1139213142e762d316a3eb2f3fcb52e8e0bdfe59