Lazy class loading with resource providers.
Extending the current lazy loading implementation to allow API
consumer to pass instances implementing LazyResourceProvider
interface as a source of resources that can be queried lazily
for on-demand class searches.

Only class files can be added this way in this implementation.

This functionality is currently implemented as an *extension*
of regular lazy loading which is based on DexClassPromise and
LazyClassFileLoader, and reuses this implementation a lot.

The advantage of using lazy resource providers is that it does
not require actually having all the resources preloaded up front,
but it comes at cost of minor differences of how errors for
conflicting classes are reported in resource-provider based
and regular scenarios.

Currently all the resources added to D8 via --lib or --classpath
are still loaded via regular lazy loading mechanism, resource
providers are only available for API users.

BUG=

Change-Id: I70cba6f3603bfd1e8e33a168d5fd8a3000a6e762