Do not workaround javac bug that leads to wrong parameter annotations.
ASM internally corrects the number of parameter annotations when javac
gets it wrong. It does so by using a non-existing java.lang.Synthetic
annotation. In the MethodWriter it then undoes that again so that the
output will maintain the javac bug. Since we are not using the
MethodWriter, we get only half of the workaround, namely the bit that
adds these fake annotations. We therefore have to filter them out
ourselves to get the right (wrong) behavior. :)

R=zerny@google.com

Bug: 62300145
Change-Id: Ieec7d165d8b2a6eb664f965693de76801e691bb8