Enable test without runtime to run on Windows
This is a first attempt to run R8 tests on windows
until support is fully implemented to use
device and emulators in place of host runtimes.

Tests which require ART are skipped (about 300 out of 2000).

As for today, some smali tests fail (11) due to the
lack of port of the dexmerger script, which prevents
from generating the dex files the tests rely on.

Bug:

Change-Id: Ia558189605f677cd6ce7ab71c24155f4a7e15ea8