Introduce fewer argument moves for ranged invokes.
When a ranged invoke is just forwarding a sub-sequence of the incoming
argument registers there is no need to introduce moves that might
end up being materialized. Instead we can use the original argument
registers. This is the first patch set of

https://r8-review.googlesource.com/c/1360/

with the addition of special handling for getting the allocated
register for ranged invokes.

Gives around 10kB savings on dexing the gmscore v10 deploy jar.

R=mikaelpeltier@google.com

Change-Id: I6ee816cb2ffffdaadd44f4af2b2ba18d9ebadd14