Use control-flow graph to compute JarState for all blocks.
This uses a fixed-point algorithm to make sure that we merge the state
from all control-flow edges to a basic block. This ensures that if
we have a null type on one edge and something more specific on another
edge, we record the specific information.

The only case where this makes a difference is for loads from
byte and boolean arrays which we have to be able to distinguish
in order to generate the right dex instruction.

R=zerny@google.com

Bug: 62438166
Change-Id: Id8665893e0f4509924e8040a1c4e3f9c4e342449