Make it less likely that Art lock verification gets confused.
When locking on argument values, use the argument register for all lock
operations whenever possible.

In all but the most obscure cases that should be possible. However,
it is not bullet-proof and when the argument registers become too high
we will have to use another register for it and lock verification
will get confused by the splitting and print a warning. That shouldn't
be treated as a fatal error, but we can attempt to avoid it as much
as possible.

R=acleung@google.com

Bug: 62475512
Change-Id: Ia4358e38aacfb90e79167d4b5144afa80d9cc30b