Make the run_on_app.py script dertimine additional output files
The output path of the Proguard map and Proguard seeds flies are
specified in the proguard configuration file. Change the script
tools/run_on_app.py to generate a temporary Proguard configuration
to direct the these two files to the same output location as the
dex files no matter what is in the input Proguard configuration.

R=ager@google.com, ricow@google.com

Change-Id: I27f27842e9a3a370dcb1f4deff91154b8d044e07