Move assertion to where it actually must hold.
When merging stack we can have a basic type on one path and the null
marker on the other path. In that case we take the basic type and
it is *not* true that the values are compatible. Only if none of
the stack values are null can is it guaranteed that they are
compatible.

R=sgjesse@google.com, zerny@google.com

Change-Id: I7232ee278102a0624323f744b047e626b480a38f