Ensure deterministic IR optimization
When the IR processing have to break cycles in the call graph to produce a
set of leaves to process, the processing order can change the generated code
as optimization feedback from processing one method feeds into another method.

To get deterministic behaviour the optimization feedback is collected without
making it available to other methods immediately. When all methods in the
set of leaves have been processed the collected optimization feedback is
applied.

After the optimization feedback has been applied the methods calling any
of the methods with new optimization feedback will be re-processed.

When methods are re-processed for other reasons (double-inlining and
outlining) no additional optimization feedback is collected to ensure
deterministic behaviour.

This also brings back concurrent IR processing as the issue with
deterministic processing was also hit there.

Bug: 37133161

Change-Id: Idf0ba09510038a0634cc1e34e44b50c6918b08cc