Update how register candidate is chosen
- Firstly try to find a register candidate that is a constant that
can be rematerialized. If it does not exist or if hos live range is
too short, fallback to select a register that can not be rematerialized.
It allows to save 14608 bytes on GMSCore V10(deploy jar), 15080 bytes
on GMSCore v9 (deploy jar) and for dex file related to the bug, it
allows to have 49880 bytes instead of 56380 bytes in debug mode and it
allows to have 48424 bytes instead of 55324 in release mode.

BUG: 62475297

Change-Id: I497c13a4c3dd8bdd6acafc6e2a3197648604035f