Use ranged invokes when out args are in args in high registers.
This allows us to use high registers for all invokes that are just
passing along incoming arguments in the right order. Using
ranged invokes we do not have to move the incoming arguments to
low registers for the call.

This saves 38kB on debug mode dexing of gmscore v10 deploy jar.

R=mikaelpeltier@google.com

Change-Id: I0c53c451636a57c7b78c58f67babe2200fb1568c