Change register allocator to deal better with unconstrained uses.
When low on registers, if the interval we are currently attempting
to allocate a register for has an unconstrained first use, split
the current interval instead of finding another interval to split.

In addition, make the determination of when to split argument
values right after their definition more precise.

These changes shouldn't matter much at this point. However,
zerny@ is working on changes to debug information handling
that will make this occur a lot more for arguments. For such
cases this is needed and generates better code than we currently
do.

R=zerny@google.com

Change-Id: I6d83b91a15486d294c2c66f68b17cb871c0cbae1