Generate better code for ranged invokes in debug mode.
The register allocator inserts moves for arguments to ranged invokes
in order to give the register allocator complete freedom in selecting
registers for those arguments.

However, debug instructions are also moves and therefore, we often
don't have to generate extra argument moves in order to have enough
freedom in register selection.

This change avoids generating argument moves when the input is
already the result of a move to an unconstrained value that can
be linked into the current argument chain.

This saves around 40kB on GMSCore v10 deploy jar dexing.

R=mikaelpeltier@google.com, zerny@google.com

Bug: 62992398
Change-Id: I60f89d06e25c4347c3a688ed65fd99b75e477986