Fixing synchronization issue with getting loaded classes from ClassMap.
ClassMap with non-null class provider may add new values to `classes`
collection for types which are known to not exist, so class map does
not have to reach out to class provider next time the same type is
asked for. This CL fixes an assertion which was triggered in such cases.

Bug:
Change-Id: I93e0418da604184dfaaa7b99e097c0d529958a7c