Don't consider library methods for single invoke targets
Methods from the library can have implementations which are not visible -
even if the library is included in the compilation. This could lead to
a single implementation of a library interface method in the application
to be considered a single target for optimization purposes.

Keep the special handling of selected library methods from
java.lang.StringBuilder and java.lang.StringBuffer.

This saves 28 KB on GMS Core v9.

Change-Id: I788186f10513bfe3644515a2a649a1d45bc1a721