Do not shared too much constants in dominator block of usages
- It does not impact dex file size (only few bytes) of deploy.jar of GMS core v10
and v9. It allows to decrease code size of generated dex of jar files from the
Android tree. Code size is decrease from 1783512 bytes to 466104 bytes for
cssparser-0.9.13.jar and from 760244 bytes to 689728 bytes for freemarker-2.3.16.jar.
In both case, the generated dex becomes smaller than with Dx.
For the jar file attached to the bug, the generated dex file is 41096 bytes
instead of 48424 bytes in release mode and is 43936 bytes instead of 49880 bytes
in debug mode. In this case, generated dex files are also smaller than with Dx.

Bug: 62475297
Change-Id: Idbab42a42c244bcc7f67e00238d38185405e8e72