Do not obfuscate unknown types in protos.
We used to obfuscate the name of classes in protos if we did not
have a definition. The idea being that the program can only work if
the corresponding parameters only get null values assigned.

This unfortunately does not work for gmscore due to its dependencies
on different android SDK versions. We end up in a scenario where a
class from KitKat library is referenced but our code does not know,
as the class is missing, and treats the class as a program class.

This costs us about 10k.

R=bak@google.com

Bug:
Change-Id: I05b43845c14a2e12dc361de16649716b16ecb746