Fix register allocation issues caused by special debug uses.
Debug uses registered on live intervals were sometimes ignored
and other times not. That leads to inconsistencies in the
register allocator.

With this change, debug uses only extend the live range of values
but do not register live interval uses. That makes the handling
of uses consistent for the register allocator again and still
ensures that values are kept alive by debug users in some
register.

R=herhut@google.com, mikaelpeltier@google.com, sgjesse@google.com, zerny@google.com

Bug: 63285632
Bug: 62992398
Change-Id: Ia6515fed81d4d0e23c25631681076a67217a6e75