Fix potential issues related to CFG printing
This CL addresses 2 complaints from FindBugs related to close of
resources and charset encoding.

The 1st issue is about the writer not being closed if an exception
happens during the write operation. This is addressed by the usage
of a try-with-resources.

The 2nd issue is about relying on the platform's default encoding
(executing the program). This is addressed by explicitly setting the
charset to UTF-8 (which requires to use a different API).

Change-Id: Iad5070e1b3a13d259b7ab7ebc28e3303fa2aabdf