Different main-dex error messages for D8/R8.
Depending on which tool is used, main-dex error messages will suggest
different uses of a main-dex list and/or main dex rules.

Also, correct the number of methods and/or fields that exceed the limit
by collecting them all. (Previously, the logic failed earlier, and thus
showed just MAX + 1, not the exact number of methods/fields in it.)

This is a follow-up of http://go/r8g/3300.

Bug: 63170711
Change-Id: I4a619353ac951ebef74e200a4ea43ca60ee371ac