Reduce memory use for DexString canonicalization when merging.
When merging we are coming from DexStrings and allocating a Java
String and keeping both the DexString and the Java String alive
in the canonicalization map is a significant overhead.

This change goes back to canonicalizing only based on DexStrings.
It keeps the speed advantage of the old code by moving an
additional map int the JarApplicationReader to provide fast lookup
based on a Java String without going through intermediate DexString
allocations.

R=sgjesse@google.com, zerny@google.com

Change-Id: I928cc38905bcb5c27fa36780f18b49a99bf324b5