Decrease peak memory usage for the common case.
Byte arrays created for files in input archives should not be reachable
from the outside again and therefore we can make them one-shot
resources that do not hold on to their byte arrays after you get
their stream.

Applications can be used multiple times. Therefore, this change
allows the rereading of the program archives which recreates
the content.

This makes IOExceptions happen at a slightly different time in the
pipeline, not sure that we care much.

R=zerny@google.com

Change-Id: Ia9b7eb78b0ea9814d37a3611180543796c31ec26