Make minimal main-dex dependent on compilation mode
Both D8 and R8 will now produce minimal main-dex based on the
compilation mode. In debug mode minimal main-dex is always produced,
so that the validity of the main-dex can be debugged. For release mode
minimal main-dex is not produced - the primary dex file will be filled
as appropriate.

This also removes the explicit minimal main-dex option from R8 by
removing the R8 API

  setMinimalMainDex(boolean value)

and the related --minimal-main-dex option.

Change-Id: I2c9ed774fe7893c9885733ae2f24e3cb195f21dd