Add Marker when running D8 without a main-dex list as well.
With detailed assertion message for MainDexListTest#deterministicTest,
currently case primary: DEBUG: without a main-dex list is failing, and
the main diff is that that case is using MonoDexDistributor whereas
other cases are landed on FillFilesDistributor. And, the diff between
two distributors is adding tool marker or not.

Desclaimer: There are two more kinds of distributors in VirtualFile,
but I am not sure whether the remaining also need to add tool marker.

Bug:
Change-Id: I230f0ccaa29d42d2fb6b8db715c54a34f3d7b0d2