Add support flattening package hierarchy.
Based on http://go/r8-pkg-obfuscation, this CL implements flattening
(as well as repackaging) packages (in a slightly different way).

* Repackaging
In ClassNameMinifier constructor, a top-level state is introduced.
That top-level state's target package prefix is user-given.
Then, when searching for a state for any can-be-renamed classes,
that top-level state is returned, hence all those classes share the same
name space.

* Flattening
The only difference is, for different packages, different states are
created and mapped whose package prefix is suggested by the top-level
state.  That is, classes in each pacakge are still in different name
spaces, while those packages share the same top-level name space.

Bug: 37764746
Change-Id: I37246fecbbbaa38a85cd6701d368072beead478b