Replace clinit static puts with initial values for static fields
when a static put in clinit is dominated by both the exit block, and all
other usages of the static field, then static put can be removed, and the
dex file data section (static_values_off) holding the initial value for
static fields can be used for setting the initial value.

Include <clinit> as a callee in the call graph when a class is mentioned.

Also fix a bug in the reading for initial values for static fields
from dex files.

Change-Id: I9d92c8de414d935eda8adff6a116d540de5d67dc