Allow Kotlin generated bytecode to be compiled.
The Kotlin compiler inserts locals information for locals that are
unused in the code. This change allows us to compile the code.
It does *not* insert instructions that allocate a register for
these unused locals. Debug locals information is not emitted
at the end for these unused locals.

We should not let debug information influende our code generation.
We cannot trust debug information as it could be completely
invalid. We need to figure out how these variables are used for
Kotlin debugging and see what we can do to support full Kotlin
debugging.

R=shertz@google.com, zerny@google.com

Change-Id: I07446e96227e702d088433e69a7e0ec653e91759