Implement package name obfuscation (w/ access modification).
The key idea is to map each sub-package prefix to its own NamingState,
which corresponds to a new name space. Such mappings are generated in
a recursive manner when visiting a class that is not in a keep list.
E.g., when visiting com.ggl.r8.A, prefixes com.ggl.r8, com.ggl, and com
are assigned to new NamingStates, and their renamed package names are
suggested by super-package's sate (superState in the code).

Bug: 36799686
Change-Id: I1968aad912d0c8d1675966a606023093d1d6827e